EMS-651 rename and restructure e2e test for broker
